+++ src/functions/geoJSON.js
@@ -7,6 +7,10 @@ export const geoJSONProps = {
     type: [Object, Array],
     custom: true,
   },
+  optionsStyle: {
+    type: [Object, Function],
+    custom: true,
+  },
 };
 
 export const setupGeoJSON = (props, leafletRef, context) => {
@@ -17,12 +21,18 @@ export const setupGeoJSON = (props, leafletRef, context) => {
   );
 
   const options = propsToLeafletOptions(props, geoJSONProps, layerOptions);
+  if (Object.prototype.hasOwnProperty.call(props, "optionsStyle")) {
+    options.style = props.optionsStyle;
+  }
 
   const methods = {
     ...layerGroupMethods,
     setGeojson(newVal) {
       leafletRef.value.clearLayers();
       leafletRef.value.addData(newVal);
+    },
+    setOptionsStyle(newVal) {
+      leafletRef.value.setStyle(newVal);
     },
     getGeoJSONData() {
       return leafletRef.value.toGeoJSON();